\name{LabTestsRprgm-package}
\alias{LabTestsRprgm-package}
\alias{LabTestsRprgm}
\docType{package}
\title{
Package for automated testing of computer labs in the course R programing.
}
\description{
Functions for automated testing computer labs in thecourse R programming.
}
\details{
\tabular{ll}{
Package: \tab LabTestsRprgm\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-02-12\cr
License: \tab LGPL \cr
}
For students: Use the function testLab("<the function to be tested>") to test.

}
\author{
Måns Magnusson
Maintainer: Måns Magnusson <mans.magnusson@liu.se>
}
\references{
See the course github page: 
}

\keyword{  }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[https://github.com/MansMeg/KursRprgm]{<pkg>}} ~~
}
