\name{markAssignment}
\alias{markAssignment}
\title{markAssignment}
\usage{
  markAssignment(assignments, filePath = file.choose(),
    cache = FALSE)
}
\arguments{
  \item{assignments}{The assignment(s) to mark / check. Can
  be a vector of multiple assignments.}

  \item{filePath}{The path to the file with the suggested
  solution.}

  \item{cache}{Should testfiles be stored locally
  temporarily (for slow internet connections).}
}
\description{
  Check / mark given assignment(s) for errors.
}
\details{
  The function will mark the assignment given in
}

